requirements.txt
beautifulsoup4==4.8.2
blinker==1.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
configobj==5.0.6
cryptography==2.8
distro==1.4.0
entrypoints==0.3
Flask==1.1.2
Flask-Ext==0.1
Flask-MySQL==1.5.1
httplib2==0.14.0
ibm-cloud-sdk-core==1.7.3
ibm-watson==4.7.1
idna==2.10
ifaddr==0.1.6
itsdangerous==1.1.0
Jinja2==2.11.2
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.1
nose==1.3.7
oauthlib==3.1.0
packaging==20.3
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==0.10.1
PyNaCl==1.3.0
pyparsing==2.4.6
pyRFC3339==1.1
PySocks==1.7.1
python-dateutil==2.8.1
python-xlib==0.23
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
redis==3.4.1
reportlab==3.5.34
requests==2.24.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rq==1.2.2
SecretStorage==2.3.1
simplejson==3.16.0
six==1.15.0
soupsieve==1.9.5
tinycss2==1.0.2
tweepy==3.9.0
Unidecode==1.1.1
urllib3==1.25.11
wadllib==1.3.3
webencodings==0.5.1
websocket-client==0.48.0
Werkzeug==1.0.1
zeroconf==0.24.4
